personalization updated for landing page
