+++ server.js
@@ -17,6 +17,7 @@ const {authRouter} = require('./routers');
 const {weatherRouter} = require('./routers'); 
 const {spotifyRouter} = require('./routers'); 
 const {youtubeRouter} = require('./routers');
+const {updateUserRouter} = require('./routers');
 
 const app = express();
 app.use(express.json());
@@ -41,11 +42,13 @@ passport.use(jwtStrategy);
 
 //routers
 app.use('/api/users/', usersRouter);
+app.use('/api/users/', updateUserRouter);
 app.use('/api/users/playlists', playlistsRouter);
 app.use('/api/auth/', authRouter);
 app.use('/api/users/weather', weatherRouter);
 app.use('/api/users/rec', spotifyRouter);
 app.use('/api/users/youtube', youtubeRouter);
+
 
 
 //error handling
